+++ cld_gui.pl
@@ -66,7 +66,7 @@ if(-d $ENV{PAR_TEMP}."/inc/"){
 }
 $| = 1;
 
-my ($script_name,$script_version,$script_date,$script_years) = ('cld','1.4.3','2017-01-25','2013-2015');
+my ($script_name,$script_version,$script_date,$script_years) = ('cld','1.4.4','2017-02-17','2013-2015');
 
 
 ###################################################################################################################################################################################################
@@ -313,7 +313,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>80,
 																	"bowtie_version"=>"bowtie",
 																	"offtargetdb"=>"genomeDNA",
-																	"off-targets-allowed"=>5,
+																	"targets-allowed"=>5,
 																	"unspecific_leading_bases"=>6,
 																	"edit_distance_allowed"=>2,
 																	"bowtie_mode"=>"very-sensitive",
@@ -357,7 +357,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>80,
 																	"bowtie_version"=>"bowtie",
 																	"offtargetdb"=>"genomeDNA",
-																	"off-targets-allowed"=>10,
+																	"targets-allowed"=>10,
 																	"unspecific_leading_bases"=>1,
 																	"edit_distance_allowed"=>1,
 																	"bowtie_mode"=>"sensitive",
@@ -398,7 +398,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>90,
 																	"bowtie_version"=>"bowtie2",
 																	"offtargetdb"=>"gDNA",
-																	"off-targets-allowed"=>20,
+																	"targets-allowed"=>20,
 																	"unspecific_leading_bases"=>0,
 																	"edit_distance_allowed"=>0,
 																	"bowtie_mode"=>"fast",
@@ -438,7 +438,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>90,
 																	"bowtie_version"=>"bowtie",
 																	"offtargetdb"=>"gDNA",
-																	"off-targets-allowed"=>5,
+																	"targets-allowed"=>5,
 																	"unspecific_leading_bases"=>3,
 																	"edit_distance_allowed"=>2,
 																	"bowtie_mode"=>"very-sensitive",
@@ -482,7 +482,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>90,
 																	"bowtie_version"=>"bowtie",
 																	"offtargetdb"=>"gDNA",
-																	"off-targets-allowed"=>5,
+																	"targets-allowed"=>5,
 																	"unspecific_leading_bases"=>3,
 																	"edit_distance_allowed"=>2,
 																	"bowtie_mode"=>"very-sensitive",
@@ -526,7 +526,7 @@ if ($something{"GUI"}) {
 																	"max_T"=>90,
 																	"bowtie_version"=>"bowtie",
 																	"offtargetdb"=>"gDNA",
-																	"off-targets-allowed"=>5,
+																	"targets-allowed"=>5,
 																	"unspecific_leading_bases"=>3,
 																	"edit_distance_allowed"=>2,
 																	"bowtie_mode"=>"very-sensitive",
@@ -574,7 +574,7 @@ if ($something{"GUI"}) {
 				#off-target specific options										;#
 				#$something{"bowtie_version"}="bowtie"								;#
 				#$something{"offtargetdb"}="gDNA"									;#
-				$something{"off-targets-allowed"}=5									;#
+				$something{"targets-allowed"}=5									;#
 				$something{"unspecific_leading_bases"}=5							;#
 				$something{"edit_distance_allowed"}=2								;#
 				#$something{"bowtie_mode"}="very-sensitive"							;#
@@ -1330,7 +1330,7 @@ if ($something{"GUI"}) {
 					   -textvariable => \$something{"offtargetdb"}
 					)																	;
 																				   ;
-					$scl_off_targets_allowed 	= $offtarget_options_frame -> Scale(-label=>"off-targets allowed [#]:",-orient=>'h',	-length=>150,	-digit=>1,	-from=>0,	-to=>100,	-variable=>\$something{"off-targets-allowed"}, -tickinterval=>33,	-resolution=>1	);	#make slider with a minimum, a maximum,an orientation, a modifed variable, a certain bckground color, the intervals ticks should be drawn, the number by that the interval should be devided 1..100 will be devided by ten so that every tenth value can be selected
+					$scl_off_targets_allowed 	= $offtarget_options_frame -> Scale(-label=>"targets allowed [#]:",-orient=>'h',	-length=>150,	-digit=>1,	-from=>1,	-to=>100,	-variable=>\$something{"targets-allowed"}, -tickinterval=>33,	-resolution=>1	);	#make slider with a minimum, a maximum,an orientation, a modifed variable, a certain bckground color, the intervals ticks should be drawn, the number by that the interval should be devided 1..100 will be devided by ten so that every tenth value can be selected
 					$scl_unspecific_leading_bases 	= $offtarget_options_frame -> Scale(-label=>"unspecific PAM distal bases [nt]:",-orient=>'h',	-length=>150,	-digit=>1,	-from=>0,	-to=>10,	-variable=>\$something{"unspecific_leading_bases"}, -tickinterval=>3,	-resolution=>1	);	#make slider with a minimum, a maximum,an orientation, a modifed variable, a certain bckground color, the intervals ticks should be drawn, the number by that the interval should be devided 1..100 will be devided by ten so that every tenth value can be selected
 					$scl_edit_distance_allowed 	= $offtarget_options_frame -> Scale(-label=>"mismatches allowed [nt]:",-orient=>'h',	-length=>150,	-digit=>1,	-from=>0,	-to=>5,	-variable=>\$something{"edit_distance_allowed"}, -tickinterval=>2,	-resolution=>1	);	#make slider with a minimum, a maximum,an orientation, a modifed variable, a certain bckground color, the intervals ticks should be drawn, the number by that the interval should be devided 1..100 will be devided by ten so that every tenth value can be selected
 					
@@ -1571,9 +1571,9 @@ if ($something{"GUI"}) {
 					$chk_retrieve_recomb_matrix	-> grid(-row=>4,-column=>1,-columnspan=>2,-sticky=>"nw");
 					$scl_right_homology			-> grid(-row=>5,-column=>1,-sticky=>"nw");
 					$scl_left_homology			-> grid(-row=>5,-column=>2,-sticky=>"nw");
-					$lab_knockout				-> grid(-row=>6,-column=>1,-columnspan=>2,-sticky=>"nw",-pady =>20);
-					$scl_downstream_window		-> grid(-row=>7,-column=>1,-sticky=>"nw");
-					$scl_upstream_window		-> grid(-row=>7,-column=>2,-sticky=>"nw");
+                    $scl_downstream_window		-> grid(-row=>6,-column=>1,-sticky=>"nw");
+					$scl_upstream_window		-> grid(-row=>6,-column=>2,-sticky=>"nw");                    
+					$lab_knockout				-> grid(-row=>7,-column=>1,-columnspan=>2,-sticky=>"nw",-pady =>20);					
 					$scl_number_of_CDS			-> grid(-row=>8,-column=>1,-sticky=>"nw");
 					$lab_paired					-> grid(-row=>9,-column=>1,-columnspan=>2,-sticky=>"nw",-pady =>20);
 					$scl_minspacerlength		-> grid(-row=>10,-column=>1,-sticky=>"nw");
@@ -2192,16 +2192,30 @@ sub make_mismatch_string{
                   $pos++
             }
       }
-      if ($_[2] eq "fw") {
+      if($_[3] eq "3_prime"){
+        if ($_[2] eq "fw") {
             foreach (1..$_[1]){
                   unshift @matchstring , "n";
             }
-      }
-      else {
+        }
+        else {
+              foreach (1..$_[1]){
+                    push @matchstring , "n";
+              }
+        }
+      }else{
+        if ($_[2] eq "fw") {            
             foreach (1..$_[1]){
-                  push @matchstring , "n";
+                    push @matchstring , "n";
+              }
+        }
+        else {
+              foreach (1..$_[1]){
+                  unshift @matchstring , "n";
             }
+        }
       }
+      
       
       
       return(@matchstring);
@@ -3186,8 +3200,8 @@ sub make_a_crispr_library{
                                     }
                               close $crisprs;
                               my $k = 30;
-                              if ($something{"off-targets-allowed"}>30) {
-                                $k=$something{"off-targets-allowed"}+1;
+                              if ($something{"targets-allowed"}>30) {
+                                $k=$something{"targets-allowed"}+1;
                               }                              
                               #####################################################################################################################################################################
                               #teemp_sec
@@ -3260,7 +3274,7 @@ sub make_a_crispr_library{
                                                             $direction = "fw";
                                                       }
                                                       if ( $line[0] =~ m/(\S+?_[^_]+_[^_]+)$/ig ) {
-                                                            my @matchstringo=make_mismatch_string (\$line,$something{"unspecific_leading_bases"}, $direction);
+                                                            my @matchstringo=make_mismatch_string (\$line,$something{"unspecific_leading_bases"}, $direction, $something{"PAM_location"});
                                                             my $cond=0;
                                                             if ($something{"PAM_location"} eq "3_prime") {
                                                                 if ( (  $direction eq "fw"
@@ -3415,7 +3429,7 @@ sub make_a_crispr_library{
                                                 if ( $line[0] =~ m/(\S+?_[^_]+_[^_]+)/ig ) {
                                                       $line[0] =~m/(\S+)_(\S+)_/;
                                                       my $seq = $1;
-                                                      my @matchstringo=make_mismatch_string (\$line,$something{"unspecific_leading_bases"}, $direction);
+                                                      my @matchstringo=make_mismatch_string (\$line,$something{"unspecific_leading_bases"}, $direction,$something{"PAM_location"});
                                                       my $startcoordinate=0;
                                                       my $spacer = abs($line[8]) - ((abs($line[8]) - abs($line[3] - $line[7])) * 2);
                                                       if ($something{"offtargetdb"} eq "genomicDNA") {                                                      
@@ -3530,7 +3544,7 @@ sub make_a_crispr_library{
                               if ( exists ${ ${ $CRISPR_hash{$fname} } {$key} }{"hits"} ) {
                                     $number_of_hits = scalar(split(";;",${ ${ $CRISPR_hash{$fname} } {$key} }{"hits"}));
                                     ${ ${ $CRISPR_hash{$fname} } {$key} }{"number_of_hits"} = $number_of_hits-1;
-                                    if (( $number_of_hits > $something{"off-targets-allowed"} + 2 || $number_of_hits < 1 ) ) {
+                                    if (( $number_of_hits > $something{"targets-allowed"} + 2 || $number_of_hits < 1 ) ) {
                                           $statistics{$fname}{"Number of designs excluded because they hit multiple targets or none"}++;
                                           delete $CRISPR_hash{$fname}{$key};
                                           next CRISPRHASHLOOP;